refactor(services): generic hosted-subprocess primitive (out of acp/)
The subprocess host (spawn an argv + surface its stdio) carries no ACP
knowledge — it's a service-tier primitive. Split it out of acp/ so both
the ACP one-shot path and the managed-agent control plane build on a
neutral module:

  * crate::subprocess::HostedSubprocess — the generic host behind a new
    unix-gated `subprocess-host` feature. `spawn_from_argv` (dup the
    stdio fds into DT_PIPEs) and `spawn_no_pipes` (hand back the tokio
    handles for a caller-driven tunnel) share a `spawn_child_piped` core.
  * acp::subprocess — thin bridge: AgentConfig → argv + env sanitising +
    the /{zone}/proc/{pid}/fd/{n} scheme, delegating via spawn_acp to the
    generic host.

service-acp pulls subprocess-host; managed_agent's raw path gates on it
too, so neither needs the other's feature.
