+++ rust/services/src/acp/mod.rs
@@ -16,8 +16,8 @@
 //!     `nexus.contracts.vfs_paths`. The Rust port keeps the same
 //!     conventions so a Python and Rust caller addressing the same
 //!     agent see the same files.
-//!   * [`subprocess`] (unix) — `AcpSubprocess` owns the agent CLI +
-//!     three stdio DT_PIPE registrations.
+//!   * [`subprocess`] (unix) — `spawn_acp` bridge: `AgentConfig` → argv
+//!     over the generic [`crate::subprocess::HostedSubprocess`] host.
 //!   * [`jsonrpc`] — newline-delimited JSON-RPC 2.0 client.
 //!   * [`observer`] — accumulator for `session/update` notifications.
 //!   * [`connection`] — ACP-specific request / notification routing.
