+++ rust/services/src/lib.rs
@@ -39,6 +39,14 @@
 //!                          +--- backends     (peer; never crosses to services)
 //! ```
 
+// Generic hosted-subprocess primitive: launch an argv + surface its
+// stdio as VFS DT_PIPEs. Shared by `acp` (config-driven ACP one-shot)
+// and `managed_agent` (raw control-plane tunnel); knows nothing about
+// either. Unix-only (dup(2) + stdio-pipe kernel support). `service-acp`
+// pulls it; a managed-agent build that wants the raw spawn path enables
+// it alongside `service-managed-agent`.
+#[cfg(all(unix, feature = "subprocess-host"))]
+pub(crate) mod subprocess;
 // AcpService — subprocess + ACP-over-stdio host for
 // `AgentKind::UNMANAGED` agents (claude / codex / gemini / …).
 #[cfg(feature = "service-acp")]
